requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
python-telegram-bot==12.4.2
requests==2.23.0
tabulate==0.8.6
urllib3==1.25.8